[feat] : Column 에러 해결

+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/applicant/domain/AnswerRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnswerRepository extends JpaRepository<Answer, Integer> {
+public interface AnswerRepository extends JpaRepository<Answer, Long> {
     List<Answer> findByApplicantIdIn(List<UUID> applicantIds);
 
     List<Answer> findByApplicantId(UUID applicantId);
