+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/applicant/domain/QuestionRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuestionRepository extends JpaRepository<Question, Integer> {}
+public interface QuestionRepository extends JpaRepository<Question, Long> {}
